vera-contextual-model/requirements.txt
flask[async]==2.3.2
flask_pymongo==2.3.0
flask-mongoengine==1.0.0
pyjwt==2.7.0
gunicorn==20.1.0
python-dotenv==1.0.0
pinecone-client==2.2.2
scikit-learn==1.3.0
tensorflow
tensorflow_hub==0.14.0

